+++ src/banner/FPSWebPartClass/IThisFPSWebPartClass.ts
@@ -1,7 +1,7 @@
 
 import { DisplayMode } from '@microsoft/sp-core-library';
 import { IFPSEnviro } from '../../common/interfaces/fps/IFPSEnviro';
-import { WebPartContextCopy_15_2 } from '../../common/interfaces/indexes';
+import { WebPartContextCopy_15_2 } from '../../common/interfaces/@msft/1.15.2/WebPartContext';
 import { ISitePreConfigProps } from '../../common/PropPaneHelp/preconfig/IPreConfig';
 import { IRepoLinks } from '../../components/atoms/Links/CreateLinks';
 import { IFieldPanelDesignMode } from '../../components/molecules/FieldPanel/components/IMinWPFieldPanelProps';