+++ src/pnpjs/SourceItems/getSourceItems.ts
@@ -1,10 +1,10 @@
 
-import { IHelpfullOutput, IHelpfullInput, convertHelpfullError } from '../../logic/indexes/HelpfullErrors';
+import { IHelpfullOutput, IHelpfullInput, convertHelpfullError } from '../../logic/Errors/friendly';
 
 import { fetchSourceItems } from "@mikezimm/fps-pnp2/lib/services/sp/fetch/items/fetchSourceItems";
 import { IItemsError, ISourceProps } from './Interface';
 import { check4Gulp } from '@mikezimm/fps-pnp2/lib/services/sp/CheckGulping';
-import { saveErrorToLog } from '../Logging';
+import { saveErrorToLog } from '../Logging/saveErrorToLog';
 
 /**
  * getSourceItems calls the Pnp function to get the results which returns the raw error.