+++ src/pnpjs/Users/logic/createISiteUsersInfoObject.ts
@@ -1,7 +1,7 @@
 
 import { ISiteUsersResults } from '@mikezimm/fps-pnp2/lib/services/sp/users/FetchSiteAdmins';
 
-import { IHelpfullInput, convertHelpfullError, IHelpfullOutput } from '../../../logic/indexes/HelpfullErrors';
+import { IHelpfullInput, convertHelpfullError, IHelpfullOutput } from '../../../logic/Errors/friendly';
 import { BaseErrorTrace } from '../../../PackageConst';
 import { IUser } from '../../../logic/Users/IUserInterfaces';
 import { createIUserFromUser } from './createIUserFromUser';